parser refactoring
